client: separate generic contract infrastructure from mixer-specific functionality
