+++ client/test_commands/scenario.py
@@ -5,12 +5,11 @@
 # SPDX-License-Identifier: LGPL-3.0+
 
 from zeth.core.zeth_address import ZethAddressPub
-from zeth.core.mixer_client import MixerClient, OwnershipKeyPair, \
-    joinsplit_sign, encrypt_notes, get_dummy_input_and_address, \
-    compute_h_sig, JoinsplitSigVerificationKey, MixCallDescription, \
-    ComputeHSigCB, JoinsplitSigKeyPair
+from zeth.core.mixer_client import MixCallDescription, MixParameters, MixResult, \
+    MixerClient, OwnershipKeyPair, JoinsplitSigVerificationKey, ComputeHSigCB, \
+    JoinsplitSigKeyPair, parse_mix_call, joinsplit_sign, encrypt_notes, \
+    get_dummy_input_and_address, compute_h_sig
 from zeth.core.zksnark import IZKSnarkProvider, ExtendedProof
-import zeth.core.contracts as contracts
 from zeth.core.constants import ZETH_PRIME, FIELD_CAPACITY
 import zeth.core.signing as signing
 from zeth.core.merkle_tree import MerkleTree
@@ -43,9 +42,9 @@ def dump_merkle_tree(mk_tree: List[bytes]) -> None:
 def wait_for_tx_update_mk_tree(
         zeth_client: MixerClient,
         mk_tree: MerkleTree,
-        tx_hash: str) -> contracts.MixResult:
+        tx_hash: str) -> MixResult:
     tx_receipt = zeth_client.web3.eth.waitForTransactionReceipt(tx_hash, 10000)
-    result = contracts.parse_mix_call(zeth_client.mixer_instance, tx_receipt)
+    result = parse_mix_call(zeth_client.mixer_instance, tx_receipt)
     for out_ev in result.output_events:
         mk_tree.insert(out_ev.commitment)
 
@@ -95,7 +94,7 @@ def bob_deposit(
         mk_tree: MerkleTree,
         bob_eth_address: str,
         keystore: mock.KeyStore,
-        tx_value: Optional[EtherValue] = None) -> contracts.MixResult:
+        tx_value: Optional[EtherValue] = None) -> MixResult:
     print(
         f"=== Bob deposits {BOB_DEPOSIT_ETH} ETH for himself and splits into " +
         f"note1: {BOB_SPLIT_1_ETH}ETH, note2: {BOB_SPLIT_2_ETH}ETH ===")
@@ -124,7 +123,7 @@ def bob_to_charlie(
         mk_tree: MerkleTree,
         input1: Tuple[int, ZethNote],
         bob_eth_address: str,
-        keystore: mock.KeyStore) -> contracts.MixResult:
+        keystore: mock.KeyStore) -> MixResult:
     print(
         f"=== Bob transfers {BOB_TO_CHARLIE_ETH}ETH to Charlie from his funds " +
         "on the mixer ===")
@@ -157,7 +156,7 @@ def charlie_withdraw(
         mk_tree: MerkleTree,
         input1: Tuple[int, ZethNote],
         charlie_eth_address: str,
-        keystore: mock.KeyStore) -> contracts.MixResult:
+        keystore: mock.KeyStore) -> MixResult:
     print(
         f" === Charlie withdraws {CHARLIE_WITHDRAW_ETH}ETH from his funds " +
         "on the Mixer ===")
@@ -187,7 +186,7 @@ def charlie_double_withdraw(
         mk_tree: MerkleTree,
         input1: Tuple[int, ZethNote],
         charlie_eth_address: str,
-        keystore: mock.KeyStore) -> contracts.MixResult:
+        keystore: mock.KeyStore) -> MixResult:
     """
     Charlie tries to carry out a double spending by modifying the value of the
     nullifier of the previous payment
@@ -296,7 +295,7 @@ def charlie_double_withdraw(
         ciphertexts,
         proof)
 
-    mix_params = contracts.MixParameters(
+    mix_params = MixParameters(
         proof,
         signing_keypair.vk,
         joinsplit_sig_charlie,
@@ -318,7 +317,7 @@ def charlie_corrupt_bob_deposit(
         mk_tree: MerkleTree,
         bob_eth_address: str,
         charlie_eth_address: str,
-        keystore: mock.KeyStore) -> contracts.MixResult:
+        keystore: mock.KeyStore) -> MixResult:
     """
     Charlie tries to break transaction malleability and corrupt the coins
     bob is sending in a transaction
@@ -399,7 +398,7 @@ def charlie_corrupt_bob_deposit(
             ciphertexts,
             proof)
 
-        mix_params = contracts.MixParameters(
+        mix_params = MixParameters(
             proof,
             joinsplit_keypair.vk,
             joinsplit_sig_charlie,
@@ -438,7 +437,7 @@ def charlie_corrupt_bob_deposit(
             charlie_eth_address,
             [fake_ciphertext0, fake_ciphertext1],
             proof)
-        mix_params = contracts.MixParameters(
+        mix_params = MixParameters(
             proof,
             new_joinsplit_keypair.vk,
             joinsplit_sig_charlie,
@@ -469,7 +468,7 @@ def charlie_corrupt_bob_deposit(
             bob_eth_address,
             ciphertexts,
             proof)
-        mix_params = contracts.MixParameters(
+        mix_params = MixParameters(
             proof,
             joinsplit_keypair.vk,
             joinsplit_sig_bob,
@@ -499,7 +498,7 @@ def charlie_corrupt_bob_deposit(
         bob_eth_address,
         ciphertexts,
         proof)
-    mix_params = contracts.MixParameters(
+    mix_params = MixParameters(
         proof,
         joinsplit_keypair.vk,
         joinsplit_sig_bob,
