+++ client/test_commands/test_ether_mixing.py
@@ -11,8 +11,7 @@ import zeth.core.utils
 import zeth.core.zksnark
 from zeth.core.prover_client import ProverClient
 from zeth.core.zeth_address import ZethAddressPriv
-from zeth.core.contracts import MixOutputEvents
-from zeth.core.mixer_client import MixerClient
+from zeth.core.mixer_client import MixOutputEvents, MixerClient
 from zeth.core.wallet import Wallet, ZethNoteDescription
 import test_commands.mock as mock
 import test_commands.scenario as scenario
