+++ client/zeth/core/contracts.py
@@ -5,107 +5,17 @@
 # SPDX-License-Identifier: LGPL-3.0+
 
 from __future__ import annotations
-from zeth.core.signing import SigningVerificationKey, Signature, \
-    verification_key_as_mix_parameter, verification_key_from_mix_parameter, \
-    signature_as_mix_parameter, signature_from_mix_parameter
-from zeth.core.pairing import PairingParameters
-from zeth.core.zksnark import IZKSnarkProvider, ExtendedProof
-from zeth.core.utils import EtherValue, hex_list_to_uint256_list
+from zeth.core.utils import EtherValue
 from zeth.core.constants import SOL_COMPILER_VERSION
 from web3.utils.contracts import find_matching_event_abi  # type: ignore
 from web3.utils.events import get_event_data  # type: ignore
-import json
 import solcx
-import traceback
 from typing import Dict, List, Iterator, Optional, Union, Iterable, Any
 
 # Avoid trying to read too much data into memory
 SYNC_BLOCKS_PER_BATCH = 1000
 
 Interface = Dict[str, Any]
-
-
-class MixParameters:
-    """
-    Arguments to the mix call.
-    """
-    def __init__(
-            self,
-            extended_proof: ExtendedProof,
-            signature_vk: SigningVerificationKey,
-            signature: Signature,
-            ciphertexts: List[bytes]):
-        self.extended_proof = extended_proof
-        self.signature_vk = signature_vk
-        self.signature = signature
-        self.ciphertexts = ciphertexts
-
-    @staticmethod
-    def from_json(zksnark: IZKSnarkProvider, params_json: str) -> MixParameters:
-        return MixParameters._from_json_dict(zksnark, json.loads(params_json))
-
-    def to_json(self) -> str:
-        return json.dumps(self._to_json_dict())
-
-    def _to_json_dict(self) -> Dict[str, Any]:
-        ext_proof_json = self.extended_proof.to_json_dict()
-        signature_vk_json = [
-            str(x) for x in verification_key_as_mix_parameter(self.signature_vk)]
-        signature_json = str(signature_as_mix_parameter(self.signature))
-        ciphertexts_json = [x.hex() for x in self.ciphertexts]
-        return {
-            "extended_proof": ext_proof_json,
-            "signature_vk": signature_vk_json,
-            "signature": signature_json,
-            "ciphertexts": ciphertexts_json,
-        }
-
-    @staticmethod
-    def _from_json_dict(
-            zksnark: IZKSnarkProvider,
-            json_dict: Dict[str, Any]) -> MixParameters:
-        ext_proof = ExtendedProof.from_json_dict(
-            zksnark, json_dict["extended_proof"])
-        signature_pk_param = [int(x) for x in json_dict["signature_vk"]]
-        signature_pk = verification_key_from_mix_parameter(signature_pk_param)
-        signature = signature_from_mix_parameter(int(json_dict["signature"]))
-        ciphertexts = [bytes.fromhex(x) for x in json_dict["ciphertexts"]]
-        return MixParameters(
-            ext_proof, signature_pk, signature, ciphertexts)
-
-
-class MixOutputEvents:
-    """
-    Event data for a single joinsplit output.  Holds address (in merkle tree),
-    commitment and ciphertext.
-    """
-    def __init__(
-            self, commitment: bytes, ciphertext: bytes):
-        self.commitment = commitment
-        self.ciphertext = ciphertext
-
-
-class MixResult:
-    """
-    Data structure representing the result of the mix call.
-    """
-    def __init__(
-            self,
-            new_merkle_root: bytes,
-            nullifiers: List[bytes],
-            output_events: List[MixOutputEvents]):
-        self.new_merkle_root = new_merkle_root
-        self.nullifiers = nullifiers
-        self.output_events = output_events
-
-
-def _event_args_to_mix_result(event_args: Any) -> MixResult:
-    mix_out_args = zip(event_args.commitments, event_args.ciphertexts)
-    out_events = [MixOutputEvents(c, ciph) for (c, ciph) in mix_out_args]
-    return MixResult(
-        new_merkle_root=event_args.root,
-        nullifiers=event_args.nullifiers,
-        output_events=out_events)
 
 
 class InstanceDescription:
@@ -218,134 +128,6 @@ def compile_files(files: List[str], **kwargs: Any) -> Any:
     return solcx.compile_files(files, optimize=True, **kwargs)
 
 
-def mix_parameters_as_contract_arguments(
-        zksnark: IZKSnarkProvider,
-        pp: PairingParameters,
-        mix_parameters: MixParameters) -> List[Any]:
-    """
-    Convert MixParameters to a list of eth ABI objects which can be passed to
-    the contract's mix method.
-    """
-    proof_contract_params = zksnark.proof_to_contract_parameters(
-        mix_parameters.extended_proof.proof, pp)
-    return [
-        proof_contract_params,
-        verification_key_as_mix_parameter(mix_parameters.signature_vk),
-        signature_as_mix_parameter(mix_parameters.signature),
-        hex_list_to_uint256_list(mix_parameters.extended_proof.inputs),
-        mix_parameters.ciphertexts,
-    ]
-
-
-def _create_web3_mixer_call(
-        zksnark: IZKSnarkProvider,
-        pp: PairingParameters,
-        mixer_instance: Any,
-        mix_parameters: MixParameters) -> Any:
-    mix_params_eth = mix_parameters_as_contract_arguments(
-        zksnark, pp, mix_parameters)
-    return mixer_instance.functions.mix(*mix_params_eth)
-
-
-def mix_call(
-        zksnark: IZKSnarkProvider,
-        pp: PairingParameters,
-        mixer_instance: Any,
-        mix_parameters: MixParameters,
-        sender_address: str,
-        wei_pub_value: int,
-        call_gas: int) -> bool:
-    """
-    Call the mix method (executes on the RPC host, without creating a
-    transaction). Returns True if the call succeeds.  False, otherwise.
-    """
-    mixer_call = _create_web3_mixer_call(
-        zksnark, pp, mixer_instance, mix_parameters)
-    try:
-        mixer_call.call({
-            'from': sender_address,
-            'value': wei_pub_value,
-            'gas': call_gas
-        })
-        return True
-
-    except ValueError:
-        print("error executing mix call:")
-        traceback.print_exc()
-
-    return False
-
-
-def mix(
-        web3: Any,
-        zksnark: IZKSnarkProvider,
-        pp: PairingParameters,
-        mixer_instance: Any,
-        mix_parameters: MixParameters,
-        sender_address: str,
-        sender_private_key: Optional[bytes],
-        pub_value: Optional[EtherValue],
-        call_gas: Optional[int]) -> str:
-    """
-    Create and broadcast a transaction that calls the mix method of the Mixer
-    """
-    mixer_call = _create_web3_mixer_call(
-        zksnark, pp, mixer_instance, mix_parameters)
-    tx_hash = send_contract_call(
-        web3, mixer_call, sender_address, sender_private_key, pub_value, call_gas)
-    return tx_hash.hex()
-
-
-def parse_mix_call(
-        mixer_instance: Any,
-        _tx_receipt: str) -> MixResult:
-    """
-    Get the logs data associated with this mixing
-    """
-    log_mix_filter = mixer_instance.eventFilter("LogMix", {'fromBlock': 'latest'})
-    log_mix_events = log_mix_filter.get_all_entries()
-    mix_results = [_event_args_to_mix_result(ev.args) for ev in log_mix_events]
-    return mix_results[0]
-
-
-def _next_nullifier_or_none(nullifier_iter: Iterator[bytes]) -> Optional[Any]:
-    try:
-        return next(nullifier_iter)
-    except StopIteration:
-        return None
-
-
-def get_mix_results(
-        web3: Any,
-        mixer_instance: Any,
-        start_block: int,
-        end_block: int,
-        batch_size: Optional[int] = None) -> Iterator[MixResult]:
-    """
-    Iterator for all events generated by 'mix' executions, over some block
-    range (inclusive of `end_block`). Batch eth RPC calls to avoid too many
-    calls, and holding huge lists of events in memory.
-    """
-    contract_address = mixer_instance.address
-    event_abi = find_matching_event_abi(mixer_instance.abi, event_name="LogMix")
-    batch_size = batch_size or SYNC_BLOCKS_PER_BATCH
-
-    while start_block <= end_block:
-        # Filters are *inclusive* wrt "toBlock", hence the -1 here, and + 1 to
-        # set start_block before iterating.
-        to_block = min(start_block + batch_size - 1, end_block)
-        filter_params = {
-            'fromBlock': start_block,
-            'toBlock': to_block,
-            'address': contract_address,
-        }
-        logs = web3.eth.getLogs(filter_params)
-        for log in logs:
-            event_data = get_event_data(event_abi, log)
-            yield _event_args_to_mix_result(event_data.args)
-        start_block = to_block + 1
-
-
 def send_contract_call(
         web3: Any,
         call: Any,
@@ -375,3 +157,34 @@ def send_contract_call(
 
     # Hosted path
     return call.transact(tx_desc)
+
+
+def get_event_logs(
+        web3: Any,
+        instance: Any,
+        event_name: str,
+        start_block: int,
+        end_block: int,
+        batch_size: Optional[int]) -> Iterator[Any]:
+    """
+    Query the attached node for all events emitted by the given contract
+    instance, with the given name. Yields an iterator of event-specific objects
+    to be decoded by the caller.
+    """
+    contract_address = instance.address
+    event_abi = find_matching_event_abi(instance.abi, event_name=event_name)
+    batch_size = batch_size or SYNC_BLOCKS_PER_BATCH
+
+    while start_block <= end_block:
+        # Filters are *inclusive* wrt "toBlock", hence the -1 here, and +1 to
+        # set start_block before iterating.
+        to_block = min(start_block + batch_size - 1, end_block)
+        filter_params = {
+            'fromBlock': start_block,
+            'toBlock': to_block,
+            'address': contract_address,
+        }
+        logs = web3.eth.getLogs(filter_params)
+        for log in logs:
+            yield get_event_data(event_abi, log)
+        start_block = to_block + 1
