+++ client/zeth/core/mixer_client.py
@@ -28,7 +28,8 @@ import os
 import json
 from Crypto import Random
 from hashlib import blake2s, sha256
-from typing import Tuple, Dict, List, Callable, Optional, Any
+import traceback
+from typing import Tuple, Dict, List, Iterator, Callable, Optional, Any
 
 
 ZERO_UNITS_HEX = "0000000000000000"
@@ -53,17 +54,6 @@ JoinsplitSigKeyPair = signing.SigningKeyPair
 
 
 ComputeHSigCB = Callable[[bytes, bytes, JoinsplitSigVerificationKey], bytes]
-
-
-class JoinsplitInputNote:
-    """
-    A ZethNote, along with the nullifier and location in Merkle tree.
-    """
-
-    def __init__(self, note: ZethNote, nullifier: str, merkle_location: int):
-        self.note = note
-        self.nullifier = nullifier
-        self.merkle_location = merkle_location
 
 
 class MixCallDescription:
@@ -112,6 +102,114 @@ class MixCallDescription:
                  for _ in range(constants.JS_OUTPUTS - len(outputs))]
         else:
             self.outputs = outputs
+
+
+class MixParameters:
+    """
+    All data required to call the mixer, with no further processing required
+    (except creation of a transaction). This is the result of fully processing
+    a MixCallDescription, generating appropriate secret data and ZK-proof and
+    signing the result.
+    """
+    def __init__(
+            self,
+            extended_proof: ExtendedProof,
+            signature_vk: signing.SigningVerificationKey,
+            signature: signing.Signature,
+            ciphertexts: List[bytes]):
+        self.extended_proof = extended_proof
+        self.signature_vk = signature_vk
+        self.signature = signature
+        self.ciphertexts = ciphertexts
+
+    @staticmethod
+    def from_json(zksnark: IZKSnarkProvider, params_json: str) -> MixParameters:
+        return MixParameters._from_json_dict(zksnark, json.loads(params_json))
+
+    def to_json(self) -> str:
+        return json.dumps(self._to_json_dict())
+
+    def _to_json_dict(self) -> Dict[str, Any]:
+        ext_proof_json = self.extended_proof.to_json_dict()
+        signature_vk_json = [
+            str(x) for x in
+            signing.verification_key_as_mix_parameter(self.signature_vk)]
+        signature_json = str(signing.signature_as_mix_parameter(self.signature))
+        ciphertexts_json = [x.hex() for x in self.ciphertexts]
+        return {
+            "extended_proof": ext_proof_json,
+            "signature_vk": signature_vk_json,
+            "signature": signature_json,
+            "ciphertexts": ciphertexts_json,
+        }
+
+    @staticmethod
+    def _from_json_dict(
+            zksnark: IZKSnarkProvider,
+            json_dict: Dict[str, Any]) -> MixParameters:
+        ext_proof = ExtendedProof.from_json_dict(
+            zksnark, json_dict["extended_proof"])
+        signature_pk_param = [int(x) for x in json_dict["signature_vk"]]
+        signature_pk = signing.verification_key_from_mix_parameter(
+            signature_pk_param)
+        signature = signing.signature_from_mix_parameter(
+            int(json_dict["signature"]))
+        ciphertexts = [bytes.fromhex(x) for x in json_dict["ciphertexts"]]
+        return MixParameters(
+            ext_proof, signature_pk, signature, ciphertexts)
+
+
+def mix_parameters_as_contract_arguments(
+        zksnark: IZKSnarkProvider,
+        pp: PairingParameters,
+        mix_parameters: MixParameters) -> List[Any]:
+    """
+    Convert MixParameters to a list of eth ABI objects which can be passed to
+    the contract's mix method.
+    """
+    proof_contract_params = zksnark.proof_to_contract_parameters(
+        mix_parameters.extended_proof.proof, pp)
+    return [
+        proof_contract_params,
+        signing.verification_key_as_mix_parameter(mix_parameters.signature_vk),
+        signing.signature_as_mix_parameter(mix_parameters.signature),
+        hex_list_to_uint256_list(mix_parameters.extended_proof.inputs),
+        mix_parameters.ciphertexts,
+    ]
+
+
+class MixOutputEvents:
+    """
+    Event data for a single joinsplit output.  Holds address (in merkle tree),
+    commitment and ciphertext.
+    """
+    def __init__(
+            self, commitment: bytes, ciphertext: bytes):
+        self.commitment = commitment
+        self.ciphertext = ciphertext
+
+
+class MixResult:
+    """
+    Data structure representing the result of the mix call.
+    """
+    def __init__(
+            self,
+            new_merkle_root: bytes,
+            nullifiers: List[bytes],
+            output_events: List[MixOutputEvents]):
+        self.new_merkle_root = new_merkle_root
+        self.nullifiers = nullifiers
+        self.output_events = output_events
+
+
+def _event_args_to_mix_result(event_args: Any) -> MixResult:
+    mix_out_args = zip(event_args.commitments, event_args.ciphertexts)
+    out_events = [MixOutputEvents(c, ciph) for (c, ciph) in mix_out_args]
+    return MixResult(
+        new_merkle_root=event_args.root,
+        nullifiers=event_args.nullifiers,
+        output_events=out_events)
 
 
 def create_zeth_notes(
@@ -382,7 +480,7 @@ class MixerClient:
         prover_config = self._prover_client.get_configuration()
         zksnark = get_zksnark_provider(prover_config.zksnark_name)
         pp = prover_config.pairing_parameters
-        return contracts.mix(
+        return mix(
             self.web3,
             zksnark,
             pp,
@@ -478,7 +576,7 @@ class MixerClient:
             signing_keypair: signing.SigningKeyPair,
             ext_proof: ExtendedProof,
             sender_eth_address: str
-    ) -> contracts.MixParameters:
+    ) -> MixParameters:
 
         # Encrypt the notes
         outputs_and_notes = zip(mix_call_desc.outputs, prover_inputs.js_outputs) \
@@ -499,11 +597,8 @@ class MixerClient:
             ciphertexts,
             ext_proof)
 
-        mix_params = contracts.MixParameters(
-            ext_proof,
-            signing_keypair.vk,
-            signature,
-            ciphertexts)
+        mix_params = MixParameters(
+            ext_proof, signing_keypair.vk, signature, ciphertexts)
         return mix_params
 
     def create_mix_parameters_and_signing_key(
@@ -516,7 +611,7 @@ class MixerClient:
             v_in: EtherValue,
             v_out: EtherValue,
             compute_h_sig_cb: Optional[ComputeHSigCB] = None
-    ) -> Tuple[contracts.MixParameters, JoinsplitSigKeyPair]:
+    ) -> Tuple[MixParameters, JoinsplitSigKeyPair]:
         """
         Convenient around creation of MixCallDescription, ProofInputs, Proof and
         MixParameters.
@@ -552,14 +647,14 @@ class MixerClient:
 
     def mix(
             self,
-            mix_params: contracts.MixParameters,
+            mix_params: MixParameters,
             sender_eth_address: str,
             sender_eth_private_key: Optional[bytes],
             tx_value: Optional[EtherValue] = None,
             call_gas: int = constants.DEFAULT_MIX_GAS_WEI) -> str:
         prover_config = self._prover_client.get_configuration()
         zksnark = get_zksnark_provider(prover_config.zksnark_name)
-        return contracts.mix(
+        return mix(
             self.web3,
             zksnark,
             prover_config.pairing_parameters,
@@ -572,13 +667,13 @@ class MixerClient:
 
     def mix_call(
             self,
-            mix_params: contracts.MixParameters,
+            mix_params: MixParameters,
             sender_eth_address: str,
             wei_pub_value: int,
             call_gas: int) -> bool:
         prover_config = self._prover_client.get_configuration()
         zksnark = get_zksnark_provider(prover_config.zksnark_name)
-        return contracts.mix_call(
+        return mix_call(
             zksnark,
             prover_config.pairing_parameters,
             self.mixer_instance,
@@ -606,7 +701,7 @@ def encrypt_notes(
 
 
 def receive_note(
-        out_ev: contracts.MixOutputEvents,
+        out_ev: MixOutputEvents,
         receiver_k_sk: EncryptionSecretKey
 ) -> Optional[Tuple[bytes, ZethNote]]:
     """
@@ -741,3 +836,98 @@ def _phi_randomness() -> bytes:
     Compute the transaction randomness "phi", used for computing the new rhoS
     """
     return bytes(Random.get_random_bytes(constants.PHI_LENGTH_BYTES))
+
+
+def _create_web3_mixer_call(
+        zksnark: IZKSnarkProvider,
+        pp: PairingParameters,
+        mixer_instance: Any,
+        mix_parameters: MixParameters) -> Any:
+    mix_params_eth = mix_parameters_as_contract_arguments(
+        zksnark, pp, mix_parameters)
+    return mixer_instance.functions.mix(*mix_params_eth)
+
+
+def mix_call(
+        zksnark: IZKSnarkProvider,
+        pp: PairingParameters,
+        mixer_instance: Any,
+        mix_parameters: MixParameters,
+        sender_address: str,
+        wei_pub_value: int,
+        call_gas: int) -> bool:
+    """
+    Call the mix method (executes on the RPC host, without creating a
+    transaction). Returns True if the call succeeds.  False, otherwise.
+    """
+    mixer_call = _create_web3_mixer_call(
+        zksnark, pp, mixer_instance, mix_parameters)
+    try:
+        mixer_call.call({
+            'from': sender_address,
+            'value': wei_pub_value,
+            'gas': call_gas
+        })
+        return True
+
+    except ValueError:
+        print("error executing mix call:")
+        traceback.print_exc()
+
+    return False
+
+
+def mix(
+        web3: Any,
+        zksnark: IZKSnarkProvider,
+        pp: PairingParameters,
+        mixer_instance: Any,
+        mix_parameters: MixParameters,
+        sender_address: str,
+        sender_private_key: Optional[bytes],
+        pub_value: Optional[EtherValue],
+        call_gas: Optional[int]) -> str:
+    """
+    Create and broadcast a transaction that calls the mix method of the Mixer
+    """
+    mixer_call = _create_web3_mixer_call(
+        zksnark, pp, mixer_instance, mix_parameters)
+    tx_hash = contracts.send_contract_call(
+        web3, mixer_call, sender_address, sender_private_key, pub_value, call_gas)
+    return tx_hash.hex()
+
+
+def parse_mix_call(
+        mixer_instance: Any,
+        _tx_receipt: str) -> MixResult:
+    """
+    Get the logs data associated with this mixing
+    """
+    log_mix_filter = mixer_instance.eventFilter("LogMix", {'fromBlock': 'latest'})
+    log_mix_events = log_mix_filter.get_all_entries()
+    mix_results = [_event_args_to_mix_result(ev.args) for ev in log_mix_events]
+    return mix_results[0]
+
+
+def _next_nullifier_or_none(nullifier_iter: Iterator[bytes]) -> Optional[Any]:
+    try:
+        return next(nullifier_iter)
+    except StopIteration:
+        return None
+
+
+def get_mix_results(
+        web3: Any,
+        mixer_instance: Any,
+        start_block: int,
+        end_block: int,
+        batch_size: Optional[int] = None) -> Iterator[MixResult]:
+    """
+    Iterator for all events generated by 'mix' executions, over some block
+    range (inclusive of `end_block`). Batch eth RPC calls to avoid too many
+    calls, and holding huge lists of events in memory.
+    """
+    logs = contracts.get_event_logs(
+        web3, mixer_instance, "LogMix", start_block, end_block, batch_size)
+    for event_data in logs:
+        yield _event_args_to_mix_result(event_data.args)
